+++ example/adaptive_style_transfer/utils.py
@@ -1,9 +1,13 @@
-from os import listdir
-from os.path import join
-
+from __future__ import print_function
+import tensorflow as tf
+import tensorlayer as tl
 import numpy as np
+from datetime import datetime
+from os import listdir, remove
+from os.path import join
 from scipy.misc import imread, imresize
 
+PATH = '/root/even/dataset/wiki_all_images'
 
 def list_images(directory):
     images = []
@@ -43,3 +47,57 @@ def get_train_images(paths, resize_len=512, crop_height=256, crop_width=256):
     images = np.stack(images, axis=0)
 
     return images
+
+
+# Normalizes the `content_features` with scaling and offset from `style_features`.
+def AdaIN(content_features, style_features, alpha=1, epsilon=1e-5):
+
+    content_mean, content_variance = tf.nn.moments(content_features, [1, 2], keep_dims=True)
+    style_mean, style_variance = tf.nn.moments(style_features, [1, 2], keep_dims=True)
+
+    normalized_content_features = tf.nn.batch_normalization(
+        content_features, content_mean, content_variance, style_mean, tf.sqrt(style_variance), epsilon
+    )
+    normalized_content_features = alpha * normalized_content_features + (1 - alpha) * content_features
+    return normalized_content_features
+
+
+def pre_process_dataset(dir_path):
+
+    paths = tl.files.load_file_list(dir_path,  regx='\\.(jpg|jpeg|png)', keep_prefix=True)
+
+    print('\norigin files number: %d\n' % len(paths))
+
+    num_delete = 0
+
+    for path in paths:
+
+        try:
+            image = imread(path, mode='RGB')
+        except IOError:
+            num_delete += 1
+            print('Cant read this file, will delete it')
+            remove(path)
+
+        if len(image.shape) != 3 or image.shape[2] != 3:
+            num_delete += 1
+            remove(path)
+            print('\nimage.shape:', image.shape, ' Remove image <%s>\n' % path)
+        else:
+            height, width, _ = image.shape
+
+            if height < width:
+                new_height = 512
+                new_width = int(width * new_height / height)
+            else:
+                new_width = 512
+                new_height = int(height * new_width / width)
+
+            try:
+                image = imresize(image, [new_height, new_width], interp='nearest')
+            except:
+                print('Cant resize this file, will delete it')
+                num_delete += 1
+                remove(path)
+
+    print('\n\ndelete %d files! Current number of files: %d\n\n' % (num_delete, len(paths) - num_delete))
